mavenize openHAB and integrate mavenized ESH repository
This commit integrates the whole mavenized ESH repository ATM.
I would prefer to drop the ESH extensions later, because IIRC there are
some openHAB bundles that depends on an extrension.
Perhaps we could drop the ESH extensions as soon as we moved some
openHAB non framework bundles to the addons repository.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>