[gson] bump gson version to 2.4
When I tried to migrate my tplinksmarthome binding to mvn + bnd (because I thought that would be easier to get test running) I found out it didn't work because the gson version it links to is 2.3.1. Unfortunately my binding uses a feature added in 2.4. Therefor I updated the gson version to the same version as the runtime.

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>