Removed apache commons-httpclient, added commons-codec.
While migrating the neato binding I encountered a problem with commons-codec versions.
Turns out the commons-codec version included is from the rather old commons-httpclient. Since all http goes through jetty I guess this library could be removed. SoI've removed the dependency on commons-httpclient and put back an up-to-date version of commons-code.
The only use in core was in the compat1x HttpUtil. This has be replaced by calls to the smarthome HttpUtil methods.

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>