Natural language processor (NLP) with modular skill architecture
Adapted from HaBot to be independent from the HaBot card system.
Thank you Yannick Schaus for this great piece of code.

The base bundle (org.openhab.nlp) is to be used with openHAB voice services.
Coupled with speech-to-text and text-to-speech engines it allows to build
privacy-focused specialized voice assistant services.

The rest bundle (org.openhab.nlp.rest) allows to interact in a chat like form.
Just send your sentence to `/nlp/chat` and corresponding actions are performed and
and an answer is generated. Use `/nlp/greet` for an initial greeting.

Each bundle features its own readme including a
"Develop for NLP in openHAB" section for skill development.

Some of the original skills were user-interface only and are
not part of this PR. The generated `IntentInterpretation`
contains all relevant information to allow skills that are ui centric as well.
A user interface system would only need to provide a separate
rest interface and might provide additional ui-specific skills.

Signed-off-by: David Graeff <david.graeff@web.de>
Also-by: Yannick Schaus <habpanel@schaus.net>