add launch configuration to generate the model sources
This launch configuration should be executed in the distros Oomph setup
as a one time job on first installation (I assume something like
`excludedTriggers="STARTUP MANUAL"`).

If it is executed by as setup instsructions no parallel auto build will
be triggered. If (for testing) executed inside a running IDE manually,
the automatic build should be disabled and a refresh should be triggered
in front of enabling it again.

If you integrate that in the Oomph Distro setup, I assume you can use
the core setup file (not used by others then me I assume) as a template
(and remove it now).

I assume it can be placed after the import task of core and beore a
build task of all projects.

```xml
    <setupTask
        xsi:type="projects:ProjectsImportTask">
      <sourceLocator
          rootFolder="${git.clone.openhabcore.location}"
          locateNestedProjects="true"/>
    </setupTask>
    <setupTask
        xsi:type="launching:LaunchTask"
        id="launch.model.gen"
        excludedTriggers="STARTUP"
        launcher="ModelGen"/>
    <setupTask
        xsi:type="projects:ProjectsBuildTask"
        excludedTriggers="STARTUP"
        refresh="true"
        clean="true"/>
    <setupTask
        xsi:type="setup:ResourceCreationTask"
        id="lifecycle-mapping"
```

Related to: https://github.com/openhab/openhab-distro/issues/927

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>