fix unexpected behavior while using runblacklist in itests
If you would like to use the runblacklist bnd instruction in an
integration test it will not work as expected.

In my current test case it is not used at all.

This is caused because an included file overrides this instruction.

IMHO it would be better to use the "Merged Instructions" feature and use
a postfix in the included file. This way you can use the non postfixed
instruction as usual.

See: https://bnd.bndtools.org/chapters/820-instructions.html#merged-instructions

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>