Add JythonScriptEngineFactory
... w/ core and custom helper libraries! Requires #1251. Fixes
https://github.com/openhab/openhab2-addons/issues/4801.

This dramatically simplifies the installation of Jython and the Jython
core and community helper libraries.

Questions:

* Should this project go into OHC or openhab2-addons? My preference is
to keep all automation in OHC and eventually split it out into another
repo.
* Is the copyOnWriteArray needed in ScriptModuleTyeProvider? I don't
think so.
* I've unrolled the Jython jar. Is there and issue with the NOTICE file?

I first used Jython 2.7.2b2, but it no longer works with recent builds
of OH (works with S1749 though). See #1252. I was seeing the same error
with 2.7.1, so this bundle uses 2.7.0.

I have another PR that makes a custom NashornScriptEngineFactory, but
I'll wait for that one until the decisions have been made for this one.
If custom ScriptEngineFactories or to go into openhab2-addons, then
NashornScriptEngineFqactory should be moved there too, which will be
difficult to have it load by default.

Signed-off-by: Scott Rushworth <openhab@5iver.com>