Remove MagicHttpResource
For ESH it made sense to use this but nowadays these UIs are no longer part of openhab-core and the demo app used for development is in openhab-distro.
Furthermore it provides links to UIs that are no longer part of OH3, uses ESH logos and the new Default UI is also using / so it throws exceptions.

See: https://github.com/openhab/openhab-distro/pull/1084#issuecomment-596226652

Signed-off-by: Wouter Born <github@maindrain.net>