Improve token issuing & session management
- Allow the use of a support cookie to mitigate leaking
of refresh tokens stored in the browser with XSS
attacks. The flow is still OAuth2 but the client can
make the server set an http-only cookie with the
session ID, which will be checked when the refresh
token is used to get another access token.
This cookie can also be used for certain servlets
(/proxy, /icon...) without an access_token to get
the user session, if the risks of CSRF are acceptable.

- Prevent leaking exceptions to the client when the token
is invalid

- Improve logging

- Detect api_key query parameters from the swagger UI

- Improvements to the authorize page: confirm the password
when creating the first account, sanitize some user input,
improve messages

- Add logout (session removal) endpoint

- Add checks yo improve conformity to RFC 6749

- Adjust some RoleAllowed annotations

- Reduce JWS key size from RS512 to RS256

Signed-off-by: Yannick Schaus <github@schaus.net>