Add support for PKCE (RFC 7636)
Support X-OPENHAB-TOKEN headers as alternative header
for the token if the X-OPENHAB-AUTH-HEADER is set
(whatever its value). This is helpful when the
Authorization header is already used by a reverse
proxy or openHAB Cloud/myopenhab.org. It can then
set a X-OPENHAB-AUTH-HEADER cookie which must be
visible to the JavaScript (i.e. no HttpOnly) which
should then pass the token in the alternative header.

Add authenticated session list endpoint

Signed-off-by: Yannick Schaus <github@schaus.net>