Fix marketplace add-on uninstall (#3042)
Instead of using the full id (`marketplace:123456`) the id without service prefix was used to lookup installed add-ons. This is not an issue if the network is available because then the information is request from the forum. If no network connection is present it obviously fails.

This also reduces the log level in the case of connection issues. ERROR is to much and the stack trace is not necessary.

Signed-off-by: Jan N. Klug <github@klug.nrw>