[core] allow vetoing ItemCommandEvents
Prevents ItemUpdater from forwarding a vetoed command to a group,
and CommunicationManager from forwarding a vetoed command to
a linked channel _or_ AutoUpdateManager.

This encompasses several components to enable it:

 * Adding a veto method to ItemCommandEvent
 * Having CommunicationManager and ItemUpdater ignore vetoed
   command events.
 * Ensuring CommunicationManager (and thus AutoUpdateManager)
   and ItemUpdater have a lower priority to process the event
   after rules have had a chance to veto the event.
   * This involves adding a priority concept to EventHandler/
     EventSubscriber; it defaults to a priority of 0
 * Adding a way to make rules run synchronously, so that their
   veto can be processed before CommunicationManager et. al.
   without race conditions
   * This includes extending rule builder and DTO
 * Adding a VetoCommandAction module for use from rules.

Note that the event must be vetoed by another EventSubscriber
with a higher priority than ItemUpdater or AutoUpdateManager.
The intercepting subscriber must veto it synchronously. This is
safe since EventHandler has a single thread for processing events
to subscribers. That said, any rule that opts in to synchronous
processing in order to be able to veto commands _must not_
use any blocking actions, since they're now run in the EventHandler
thread.

Signed-off-by: Cody Cutrer <cody@cutrer.us>