Fix ThingManager is missing config description during normalization
Reported on the forum. In some cases the `ThingManageImpl` tried to initialize the thing before all config descriptions are loaded.

This can happen because the `OH-INF/config` directory is processed in parallel to the thing descriptions. In case the handler factory is added and the thing description parser sets the corresponding `ReadyMarker` before the config descriptions are completly parsed this results in a warning because the config description cannot be found during normalization of the thing or channel configuration.

Signed-off-by: Jan N. Klug <github@klug.nrw>